refactor z3 batfish jobs
