+++ projects/batfish/src/main/java/org/batfish/z3/NodJob.java
@@ -1,155 +1,19 @@
 package org.batfish.z3;
 
-import com.google.common.collect.ImmutableMap;
-import com.microsoft.z3.BitVecNum;
+import com.google.common.collect.ImmutableSet.Builder;
+import com.microsoft.z3.BitVecExpr;
 import com.microsoft.z3.BoolExpr;
 import com.microsoft.z3.Context;
-import com.microsoft.z3.Expr;
-import com.microsoft.z3.Fixedpoint;
-import com.microsoft.z3.FuncDecl;
-import com.microsoft.z3.Model;
-import com.microsoft.z3.Params;
-import com.microsoft.z3.Solver;
-import com.microsoft.z3.Status;
-import com.microsoft.z3.Z3Exception;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
 import java.util.SortedSet;
-import java.util.TreeSet;
-import java.util.function.Function;
-import org.batfish.common.BatfishException;
+import java.util.Map.Entry;
 import org.batfish.common.Pair;
 import org.batfish.config.Settings;
-import org.batfish.datamodel.Flow;
-import org.batfish.datamodel.Ip;
-import org.batfish.datamodel.IpProtocol;
-import org.batfish.datamodel.State;
-import org.batfish.job.BatfishJob;
 
-public final class NodJob extends BatfishJob<NodJobResult> {
-
-  public static Flow createFlow(
-      String node, String vrf, Map<HeaderField, Long> constraints, String tag) {
-    Flow.Builder flowBuilder = new Flow.Builder();
-    flowBuilder.setIngressNode(node);
-    flowBuilder.setTag(tag);
-    constraints.forEach(
-        (headerField, value) -> {
-          if (headerField instanceof BasicHeaderField) {
-            switch ((BasicHeaderField) headerField) {
-              case DST_IP:
-                flowBuilder.setDstIp(new Ip(value));
-                break;
-
-              case SRC_PORT:
-                flowBuilder.setSrcPort(value.intValue());
-                break;
-
-              case DST_PORT:
-                flowBuilder.setDstPort(value.intValue());
-                break;
-
-              case FRAGMENT_OFFSET:
-                flowBuilder.setFragmentOffset(value.intValue());
-                break;
-
-              case IP_PROTOCOL:
-                flowBuilder.setIpProtocol(IpProtocol.fromNumber(value.intValue()));
-                break;
-
-              case DSCP:
-                flowBuilder.setDscp(value.intValue());
-                break;
-
-              case ECN:
-                flowBuilder.setEcn(value.intValue());
-                break;
-
-              case STATE:
-                flowBuilder.setState(State.fromNum(value.intValue()));
-                break;
-
-              case ICMP_TYPE:
-                flowBuilder.setIcmpType(value.intValue());
-                break;
-
-              case ICMP_CODE:
-                flowBuilder.setIcmpCode(value.intValue());
-                break;
-
-              case ORIG_SRC_IP:
-                flowBuilder.setSrcIp(new Ip(value));
-                break;
-
-              case SRC_IP:
-                break;
-
-              case PACKET_LENGTH:
-                flowBuilder.setPacketLength(value.intValue());
-                break;
-
-              case TCP_FLAGS_CWR:
-                flowBuilder.setTcpFlagsCwr(value.intValue());
-                break;
-
-              case TCP_FLAGS_ECE:
-                flowBuilder.setTcpFlagsEce(value.intValue());
-                break;
-
-              case TCP_FLAGS_URG:
-                flowBuilder.setTcpFlagsUrg(value.intValue());
-                break;
-
-              case TCP_FLAGS_ACK:
-                flowBuilder.setTcpFlagsAck(value.intValue());
-                break;
-
-              case TCP_FLAGS_PSH:
-                flowBuilder.setTcpFlagsPsh(value.intValue());
-                break;
-
-              case TCP_FLAGS_RST:
-                flowBuilder.setTcpFlagsRst(value.intValue());
-                break;
-
-              case TCP_FLAGS_SYN:
-                flowBuilder.setTcpFlagsSyn(value.intValue());
-                break;
-
-              case TCP_FLAGS_FIN:
-                flowBuilder.setTcpFlagsFin(value.intValue());
-                break;
-
-              default:
-                throw new BatfishException(
-                    String.format(
-                        "Unsupported %s: %s", BasicHeaderField.class.getSimpleName(), headerField));
-            }
-          } else if (headerField instanceof TransformationHeaderField) {
-            switch ((TransformationHeaderField) headerField) {
-              default:
-                throw new BatfishException(
-                    String.format(
-                        "Unsupported %s: %s",
-                        TransformationHeaderField.class.getSimpleName(), headerField));
-            }
-          }
-        });
-    return flowBuilder.build();
-  }
+public final class NodJob extends AbstractNodJob {
 
   private Synthesizer _dataPlaneSynthesizer;
 
-  private final SortedSet<Pair<String, String>> _nodeVrfSet;
-
   private QuerySynthesizer _querySynthesizer;
-
-  private String _tag;
 
   public NodJob(
       Settings settings,
@@ -157,118 +21,19 @@ public final class NodJob extends BatfishJob<NodJobResult> {
       QuerySynthesizer querySynthesizer,
       SortedSet<Pair<String, String>> nodeVrfSet,
       String tag) {
-    super(settings);
+    super(settings, nodeVrfSet, tag);
     _dataPlaneSynthesizer = dataPlaneSynthesizer;
     _querySynthesizer = querySynthesizer;
-    _nodeVrfSet = new TreeSet<>();
-    _nodeVrfSet.addAll(nodeVrfSet);
-    _tag = tag;
   }
 
   @Override
-  public NodJobResult call() {
-    long startTime = System.currentTimeMillis();
-    long elapsedTime;
-    try (Context ctx = new Context()) {
-      ReachabilityProgram baseProgram = _dataPlaneSynthesizer.synthesizeNodDataPlaneProgram();
-      ReachabilityProgram queryProgram =
-          _querySynthesizer.getReachabilityProgram(_dataPlaneSynthesizer.getInput());
-      NodProgram program = new NodProgram(ctx, baseProgram, queryProgram);
-      //      dumpAndQuit(program);
-      Params p = ctx.mkParams();
-      p.add("timeout", _settings.getZ3timeout());
-      p.add("fixedpoint.engine", "datalog");
-      p.add("fixedpoint.datalog.default_relation", "doc");
-      p.add("fixedpoint.print_answer", true);
-      Fixedpoint fix = ctx.mkFixedpoint();
-      fix.setParameters(p);
-      for (FuncDecl relationDeclaration : program.getContext().getRelationDeclarations().values()) {
-        fix.registerRelation(relationDeclaration);
-      }
-      for (BoolExpr rule : program.getRules()) {
-        fix.addRule(rule, null);
-      }
-      for (BoolExpr query : program.getQueries()) {
-        Status status = fix.query(query);
-        switch (status) {
-          case SATISFIABLE:
-            break;
-          case UNKNOWN:
-            throw new BatfishException("Query satisfiability unknown");
-          case UNSATISFIABLE:
-            break;
-          default:
-            throw new BatfishException("invalid status");
-        }
-      }
-      Expr answer = fix.getAnswer();
-      BoolExpr solverInput;
-      if (answer.getArgs().length > 0) {
-        List<Expr> reversedVarList = new ArrayList<>();
-        reversedVarList.addAll(program.getContext().getVariablesAsConsts().values());
-        Collections.reverse(reversedVarList);
-        Expr[] reversedVars = reversedVarList.toArray(new Expr[] {});
-        Expr substitutedAnswer = answer.substituteVars(reversedVars);
-        solverInput = (BoolExpr) substitutedAnswer;
-      } else {
-        solverInput = (BoolExpr) answer;
-      }
-      if (_querySynthesizer.getNegate()) {
-        solverInput = ctx.mkNot(solverInput);
-      }
-      Solver solver = ctx.mkSolver();
-      solver.add(solverInput);
-      Status solverStatus = solver.check();
-      switch (solverStatus) {
-        case SATISFIABLE:
-          break;
-
-        case UNKNOWN:
-          throw new BatfishException("Stage 2 query satisfiability unknown");
-
-        case UNSATISFIABLE:
-          elapsedTime = System.currentTimeMillis() - startTime;
-          return new NodJobResult(elapsedTime, _logger.getHistory());
-
-        default:
-          throw new BatfishException("invalid status");
-      }
-      Model model = solver.getModel();
-      Map<HeaderField, Long> constraints =
-          Arrays.stream(model.getConstDecls())
-              .map(FuncDecl::getName)
-              .map(Object::toString)
-              .map(HeaderField::parse)
-              .collect(
-                  ImmutableMap.toImmutableMap(
-                      Function.identity(),
-                      headerField ->
-                          ((BitVecNum)
-                                  model.getConstInterp(
-                                      program
-                                          .getContext()
-                                          .getVariablesAsConsts()
-                                          .get(headerField.getName())))
-                              .getLong()));
-      Set<Flow> flows = new HashSet<>();
-      for (Pair<String, String> nodeVrf : _nodeVrfSet) {
-        String node = nodeVrf.getFirst();
-        String vrf = nodeVrf.getSecond();
-        Flow flow = createFlow(node, vrf, constraints);
-        flows.add(flow);
-      }
-      elapsedTime = System.currentTimeMillis() - startTime;
-      return new NodJobResult(elapsedTime, _logger.getHistory(), flows);
-    } catch (Z3Exception e) {
-      elapsedTime = System.currentTimeMillis() - startTime;
-      return new NodJobResult(
-          elapsedTime,
-          _logger.getHistory(),
-          new BatfishException("Error running NoD on concatenated data plane", e));
-    }
-  }
-
-  private Flow createFlow(String node, String vrf, Map<HeaderField, Long> constraints) {
-    return createFlow(node, vrf, constraints, _tag);
+  protected BoolExpr computeSmtInput(
+      long startTime, Context ctx, Builder<Entry<String, BitVecExpr>> variablesAsConstsBuilder) {
+    ReachabilityProgram baseProgram = _dataPlaneSynthesizer.synthesizeNodDataPlaneProgram();
+    ReachabilityProgram queryProgram =
+        _querySynthesizer.getReachabilityProgram(_dataPlaneSynthesizer.getInput());
+    NodProgram program = new NodProgram(ctx, baseProgram, queryProgram);
+    variablesAsConstsBuilder.addAll(program.getNodContext().getVariablesAsConsts().entrySet());
+    return computeSmtConstraintsViaNod(program, _querySynthesizer.getNegate());
   }
 }
